+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add more schema configuration support for Redis based MemoryStorage connection strings
 - Update referenced asset items on content item save
 - Add App.utils.isNumeric helper function
+- Add default icon-set
 
 ## 2.8.1 (2024-02-11)